working on moving basic math stuff to basic_math.py
